Use a macro to generate `read_all_frames` tests.
